Single BNO in main

+++ ESP-Sensor/include/bnoStruct.h
@@ -5,7 +5,7 @@ using namespace std;
 struct BNOStruct
 {
     array<int16_t, 4> quat; // Stored as W, X, Y, Z
-    array<int16_t, 3> euler; // Stored as yaw, pitch, roll as radians
+    array<float, 3> euler; // Stored as yaw, pitch, roll as radians
     array<int16_t, 3> acceleration; // Stored as X, Y, Z
     array<int16_t, 3> gyro; // Stored as X, Y, Z
     array<int16_t, 3> mag; // Stored as X, Y, Z
